Refactored layer-two functions into their own module
